fix: instrument Eq derive, integration tests, and fmt
