+++ contracts/credit/Cargo.toml
@@ -32,6 +32,11 @@ name = "budget_regression"
 path = "tests/budget_regression.rs"
 required-features = ["instrument"]
 
+[[test]]
+name = "instrument"
+path = "tests/instrument.rs"
+required-features = ["instrument"]
+
 [dev-dependencies]
 gateway-auction = { path = "../../gateway-contract/contracts/auction_contract" }
 soroban-sdk = { workspace = true, features = ["testutils"] }
